Start game

+++ src/Exceptions/NotEnoughWordsException.java
@@ -8,7 +8,7 @@ public class NotEnoughWordsException extends RuntimeException {
         return msg;
     }
 
-    public NotEnoughWordsException(long numWords, int numCards, boolean black){
+    public NotEnoughWordsException(int numWords, int numCards, boolean black){
         this.msg = "Game can't start with "+(black ? "black" :"") +numCards+
                 " cards since there are only "+numWords+(black ? "black" :"") +" words in file.";
     }
